+++ test/Resources/DecoratingMailer.js
@@ -1,0 +1,9 @@
+export default class DecoratingMailer {
+  constructor (inner) {
+    this._inner = inner
+  }
+
+  get inner () {
+    return this._inner
+  }
+}